+++ ports/cef/interfaces/cef_context_menu_handler.rs
@@ -99,7 +99,7 @@ pub struct _cef_context_menu_handler_t {
   //
   // The reference count. This will only be present for Rust instances!
   //
-  pub ref_count: uint,
+  pub ref_count: usize,
 
   //
   // Extra data. This will only be present for Rust instances!
@@ -426,7 +426,7 @@ pub struct _cef_context_menu_params_t {
   //
   // The reference count. This will only be present for Rust instances!
   //
-  pub ref_count: uint,
+  pub ref_count: usize,
 
   //
   // Extra data. This will only be present for Rust instances!
